+++ adls-plugins/src/main/java/io/cdap/plugin/sink/AbstractStructuredRecordTransformer.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright © 2017 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package io.cdap.plugin.sink;
+
+import io.cdap.cdap.api.data.format.StructuredRecord;
+import io.cdap.cdap.api.data.schema.Schema;
+import io.cdap.plugin.common.RecordConverter;
+
+import java.io.IOException;
+import javax.annotation.Nullable;
+
+/**
+ * A converter that transforms Structured Record to a given OUTPUT format
+ */
+public abstract class AbstractStructuredRecordTransformer<OUTPUT> extends RecordConverter<StructuredRecord, OUTPUT> {
+  protected final Schema outputCDAPSchema;
+
+  public AbstractStructuredRecordTransformer(@Nullable io.cdap.cdap.api.data.schema.Schema outputSchema) {
+    outputCDAPSchema = outputSchema;
+  }
+
+  public OUTPUT transform(StructuredRecord structuredRecord) throws IOException {
+    return transform(structuredRecord,
+                     outputCDAPSchema == null ? structuredRecord.getSchema() : outputCDAPSchema);
+  }
+}
